added SQLite3 Database to store metadata and make the file management strict and proper

+++ app.py
@@ -2,34 +2,34 @@ from flask import Flask, render_template, request, send_file, abort
 from pathlib import Path
 from werkzeug.utils import secure_filename
 import uuid
+import database
+from datetime import datetime
 
 app = Flask(__name__)
 
 UPLOAD_FOLDER = Path("uploads")
 UPLOAD_FOLDER.mkdir(exist_ok=True)
 
+with app.app_context():
+    database.init_db()
+
+''' Cleanup function: To register a callback, whenever the application is torn down, call database.close_db() '''
+app.teardown_appcontext(database.close_db)      
+
 @app.route("/")
 def home():
+    files = database.list_files()
 
-    files = []
-
-    for file in UPLOAD_FOLDER.iterdir():
-        if not file.is_file() or file.name.startswith("."):
-            continue
-
-        original_file = file.name.split("_",1)[1]
-
-        files.append({
-            "stored_name":file.name,
-            "original_name":original_file
-        })
-        
-    return render_template("index.html",files=files)
+    return render_template(
+        "index.html",
+        files=files
+        )
 
 @app.route("/upload", methods = ["POST"])
 def upload():
     uploaded_filenames = []
     uploaded_files = request.files.getlist("file")
+
     if not uploaded_files:
         return "No files uploaded"
     for uploaded_file in uploaded_files:
@@ -41,43 +41,56 @@ def upload():
         stored_filename = f"{unique_id}_{original_filename}"
 
         destination = UPLOAD_FOLDER / stored_filename
+
         uploaded_file.save(destination)
         uploaded_filenames.append(original_filename)
+        upload_time = datetime.now().isoformat()
+        size = destination.stat().st_size
+        
+        database.create_file(str(unique_id),
+                            original_filename,
+                            stored_filename,
+                            upload_time,
+                            size)    
 
     if not uploaded_filenames:
         return "No files uploaded."
+
     return render_template(
                             "success.html",
                             files = uploaded_filenames
                         )
 
-@app.route("/download/<filename>", methods = ["GET"])
-def download(filename):
+@app.route("/download/<stored_filename>", methods = ["GET"])
+def download(stored_filename):
 
-    destination = UPLOAD_FOLDER / filename
-
-    original_name = destination.name.split("_",1)[1]
-
-    if not destination.is_file():
+    file = database.get_file(stored_filename)
+    if not file:
         abort(404)
+
+    destination = UPLOAD_FOLDER / file["stored_filename"]
     return send_file(destination,
                      as_attachment=True,
-                     download_name=original_name
-                     )
+                     download_name=file["original_filename"]
+                    )
 
-@app.route("/delete/<filename>", methods=["POST"])
-def delete(filename):
-    destination = UPLOAD_FOLDER / filename
-    original_file = filename.split("_",1)[1]
+@app.route("/delete/<stored_filename>", methods=["POST"])
+def delete(stored_filename):
 
+    file = database.get_file(stored_filename)
+    if file is None:
+        abort(404)
+    
+    destination = UPLOAD_FOLDER / file["stored_filename"]
     if not destination.is_file():
         abort(404)
     try:
         destination.unlink()
+        database.delete_file(stored_filename)
     except Exception as e:
         return f"Deletion failed due to {e}"
     
-    return render_template("deleted.html", f = original_file)
+    return render_template("deleted.html", f = file["original_filename"])
